Recipients::SecretsController
Using a token generated during secret creation, a recipient can get to a form that allows them to enter in a password to decrypt a secret message.
Added decryption feature to Secret.
Now with SMS message and email feature support.
Test coverage for the recipient decrypt feature.
Handle invalid recipient tokens with a 404.